sdk/python/pyproject.toml
name = "emergent-sdk"
version = "0.1.0"
description = "Python SDK for Emergent API"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
python-socketio = "^5.10.0"
python-engineio = "^4.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
mypy = "^1.7.0"
